requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
cs50==4.0.3
Flask==2.0.3
Flask-Session==0.4.0
flask-talisman==0.8.1
Jinja2==3.0.3
jsbeautifier==1.14.0
lxml==4.8.0
numpy==1.22.2
pandas==1.4.1
psycopg2-binary==2.9.3
redis==4.1.4
requests==2.27.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
urllib3==1.26.8
Werkzeug==2.0.3
